Change: Refactor the package framework for more clear structure.
