+++ llyra/base/strategys.py
